+++ src/drlx/sampling/__init__.py
@@ -88,329 +88,3 @@ class Sampler:
             return denoiser_unwrapped.postprocess(latents)
         else:
             return latents
-
-class DDPOSampler(Sampler):
-    def step_and_logprobs(self,
-        scheduler,
-        pred : TensorType["b", "c", "h", "w"],
-        t : float,
-        latents : TensorType["b", "c", "h", "w"],
-        old_pred : Optional[TensorType["b", "c", "h", "w"]] = None
-    ):
-        """
-        Steps backwards using scheduler. Considers the prediction as an action sampled
-        from a normal distribution and returns average log probability for that prediction.
-        Can also be used to find probability of current model giving some other prediction (old_pred)
-
-        :param scheduler: Scheduler being used for diffusion process
-        :param pred: Denoiser prediction with CFG and scaling accounted for
-        :param t: Timestep in diffusion process
-        :param latents: Latent vector given as input to denoiser
-        :param old_pred: Alternate prediction. If given, computes log probability of current model predicting alternative output.
-        """
-        scheduler_out = scheduler.step(pred, t, latents, self.config.eta, variance_noise=0)
-        
-        # computing log_probs
-        t_1 = t - scheduler.config.num_train_timesteps // self.config.num_inference_steps
-        variance = scheduler._get_variance(t, t_1)
-        std_dev_t = self.config.eta * variance ** 0.5
-        prev_sample_mean = scheduler_out.prev_sample
-        prev_sample = prev_sample_mean + torch.randn_like(prev_sample_mean) * std_dev_t
-
-        std_dev_t = torch.clip(std_dev_t, 1e-6) # force sigma > 1e-6
-
-        # If old_pred provided, we are finding probability of new model outputting same action as before
-        # Otherwise finding probability of current action
-        action = old_pred if old_pred is not None else prev_sample # Log prob of new model giving old output
-        log_probs = -((action.detach() - prev_sample_mean) ** 2) / (2 * std_dev_t ** 2) - torch.log(std_dev_t) - math.log(math.sqrt(2 * math.pi))
-        log_probs = eo.reduce(log_probs, 'b c h w -> b', 'mean')
-
-        return prev_sample, log_probs
-
-    @torch.no_grad()
-    def sample(
-        self, prompts, denoiser, device,
-        show_progress : bool = False,
-        accelerator = None
-    ) -> Iterable[torch.Tensor]:
-        """
-        DDPO sampling is analagous to playing a game in an RL environment. This function samples
-        given denoiser and prompts but in addition to giving latents also gives log probabilities
-        for predictions as well as ALL predictions (i.e. at each timestep)
-
-        :param prompts: Text prompts to condition denoiser
-        :param denoiser: Denoising model
-        :param device: Device to do inference on
-        :param show_progress: Display progress bar?
-        :param accelerator: Accelerator object for accelerated training (optional)
-        
-        :return: triple of final denoised latents, all model predictions,  all log probabilities for each prediction
-        """
-
-        if accelerator is None:
-            denoiser_unwrapped = denoiser
-        else:
-            denoiser_unwrapped = accelerator.unwrap_model(denoiser)
-
-        scheduler = denoiser_unwrapped.scheduler
-        preprocess = denoiser_unwrapped.preprocess
-        noise_shape = denoiser_unwrapped.get_input_shape()
-
-        text_embeds = preprocess(
-            prompts, mode = "embeds", device = device,
-            num_images_per_prompt = 1,
-            do_classifier_free_guidance = self.config.guidance_scale > 1.0
-        ).detach()
-
-        scheduler.set_timesteps(self.config.num_inference_steps, device = device)
-        latents = torch.randn(len(prompts), *noise_shape, device = device)
-
-        all_step_preds, all_log_probs = [latents], []
-
-        for t in tqdm(scheduler.timesteps, disable = not show_progress):
-            latent_input = torch.cat([latents] * 2)  
-            latent_input = scheduler.scale_model_input(latent_input, t)
-
-            pred = denoiser(
-                pixel_values = latent_input,
-                time_step = t,
-                text_embeds = text_embeds
-            )
-
-            # cfg
-            pred = self.cfg_rescale(pred)
-
-            # step
-            prev_sample, log_probs = self.step_and_logprobs(scheduler, pred, t, latents)
-
-            all_step_preds.append(prev_sample)
-            all_log_probs.append(log_probs)
-            latents = prev_sample
-        
-        return latents, torch.stack(all_step_preds), torch.stack(all_log_probs)
-    
-    def compute_loss(
-        self, prompts, denoiser, device,
-        show_progress : bool = False,
-        advantages = None, old_preds = None, old_log_probs = None,
-        method_config : DDPOConfig = None,
-        accelerator = None
-    ):
-
-
-        """
-        Computes the loss for the DDPO sampling process. This function is used to train the denoiser model.
-
-        :param prompts: Text prompts to condition the denoiser
-        :param denoiser: Denoising model
-        :param device: Device to perform model inference on
-        :param show_progress: Whether to display a progress bar for the sampling steps
-        :param advantages: Normalized advantages obtained from reward computation
-        :param old_preds: Previous predictions from past model
-        :param old_log_probs: Log probabilities of predictions from past model
-        :param method_config: Configuration for the DDPO method
-        :param accelerator: Accelerator object for accelerated training (optional)
-
-        :return: Total loss computed over the sampling process
-        """
-
-        # All metrics are reduced and gathered before result is returned
-        metrics = {
-            "loss" : [],
-            "kl_div" : [], # ~ KL div between new policy and old one (average)
-            "clip_frac" : [], # Proportion of policy updates where magnitude of update was clipped
-        }
-
-        scheduler = accelerator.unwrap_model(denoiser).scheduler
-        preprocess = accelerator.unwrap_model(denoiser).preprocess
-
-        adv_clip = method_config.clip_advantages # clip value for advantages
-        pi_clip = method_config.clip_ratio # clip value for policy ratio
-
-        text_embeds = preprocess(
-            prompts, mode = "embeds", device = device,
-            num_images_per_prompt = 1,
-            do_classifier_free_guidance = self.config.guidance_scale > 1.0
-        ).detach()
-
-        scheduler.set_timesteps(self.config.num_inference_steps, device = device)
-        total_loss = 0.
-
-        for i, t in enumerate(tqdm(scheduler.timesteps, disable = not show_progress)):
-            latent_input = torch.cat([old_preds[i].detach()] * 2)
-            latent_input = scheduler.scale_model_input(latent_input, t)
-
-            pred = denoiser(
-                pixel_values = latent_input,
-                time_step = t,
-                text_embeds = text_embeds
-            )
-
-            # cfg
-            pred = self.cfg_rescale(pred)
-
-            # step 
-            prev_sample, log_probs = self.step_and_logprobs(
-                scheduler, pred, t, old_preds[i],
-                old_preds[i+1]
-            )
-
-            # Need to be computed and detached again because of autograd weirdness
-            clipped_advs = torch.clip(advantages,-adv_clip,adv_clip).detach()
-
-            # ppo actor loss
-            
-            ratio = torch.exp(log_probs - old_log_probs[i].detach())
-            surr1 = -clipped_advs * ratio
-            surr2 = -clipped_advs * torch.clip(ratio, 1. - pi_clip, 1. + pi_clip)
-            loss = torch.max(surr1, surr2).mean()
-            if accelerator is not None:
-                accelerator.backward(loss)
-            else:
-                loss.backward()
-
-            # Metric computations
-            kl_div = 0.5 * (log_probs - old_log_probs[i]).mean() ** 2
-            clip_frac = ((ratio < 1 - pi_clip) | (ratio > 1 + pi_clip)).float().mean()
-            
-            metrics["loss"].append(loss.item())
-            metrics["kl_div"].append(kl_div.item())
-            metrics["clip_frac"].append(clip_frac.item())
-
-        # Reduce across timesteps then across devices
-        for k in metrics:
-            metrics[k] = torch.tensor(metrics[k]).mean().cuda() # Needed for reduction to work
-        if accelerator is not None:
-            metrics = accelerator.reduce(metrics, 'mean')
-
-        return metrics
-
-class DPOSampler(Sampler):
-    def compute_loss(
-        self,
-        prompts,
-        chosen_img,
-        rejected_img,
-        denoiser,
-        vae,
-        device,
-        method_config,
-        accelerator = None,
-        ref_denoiser = None
-    ):
-        """
-        Compute metrics and do backwards pass on loss. Assumes LoRA if reference is not given.
-        """
-        do_lora = ref_denoiser is None
-
-        scheduler = accelerator.unwrap_model(denoiser).scheduler
-        preprocess = accelerator.unwrap_model(denoiser).preprocess
-        encode = accelerator.unwrap_model(vae).encode
-
-        beta = method_config.beta
-        ref_strategy = method_config.ref_mem_strategy
-
-        # Text and image preprocessing
-        with torch.no_grad():
-            text_embeds = preprocess(
-                prompts, mode = "embeds", device = device,
-                num_images_per_prompt = 1,
-                do_classifier_free_guidance = self.config.guidance_scale > 1.0
-            ).detach()
-
-            chosen_latent = encode(chosen_img).latent_dist.sample()
-            rejected_latent = encode(rejected_img).latent_dist.sample()
-
-        # sample random ts
-        timesteps = torch.randint(
-            0, self.config.num_inference_steps, (len(chosen_img),), device = device, dtype = torch.long
-        )
-
-        # One step of noising to samples
-        noise = torch.randn_like(chosen_latent) # [B, C, H, W]
-
-        # Doubling across chosen and rejeceted
-        def double_up(x):
-            return torch.cat([x,x], dim = 0)
-        
-        def double_down(x):
-            n = len(x)
-            return x[:n//2], x[n//2:]
-
-        # Double everything up so we can input both chosen and rejected at the same time
-        timesteps = double_up(timesteps)
-        noise = double_up(noise)
-        text_embeds = double_up(text_embeds)
-        latent = torch.cat([chosen_latent, rejected_latent])
-
-        noisy_inputs = scheduler.add_noise(
-            latent,
-            noise,
-            timesteps
-        )
-
-        # Get targets
-        if scheduler.config.prediction_type == "epsilon":
-            target = noise
-        elif scheduler.config.prediction_type == "v_prediction":
-            target = scheduler.get_velocity(
-                latent,
-                noise,
-                timesteps
-            )
-        
-        # utility function to get loss simpler
-        def split_mse(pred, target):
-            mse = eo.reduce(F.mse_loss(pred, target, reduction = 'none'), 'b ... -> b', reduction = "mean")
-            chosen, rejected = double_down(mse)
-            return chosen - rejected, mse.mean()
-
-        # Forward pass and loss for DPO denoiser
-        pred = denoiser(
-            pixel_values = noisy_inputs,
-            time_step = timesteps,
-            text_embeds = text_embeds
-        )
-        model_diff, base_loss = split_mse(pred, target)
-
-        # Forward pass and loss for refrence
-        with torch.no_grad():
-            if do_lora:
-                accelerator.unwrap_model(denoiser).disable_adapters()
-
-                ref_pred = denoiser(
-                    pixel_values = noisy_inputs,
-                    time_step = timesteps,
-                    text_embeds = text_embeds
-                )
-                ref_diff, ref_loss = split_mse(ref_pred, target)
-
-                accelerator.unwrap_model(denoiser).enable_adapters()
-            else:
-                ref_inputs = {
-                    "sample" : noisy_inputs.half() if ref_strategy == "half" else noisy_inputs,
-                    "timestep" : timesteps,
-                    "encoder_hidden_states" : text_embeds.half() if ref_strategy == "half" else text_embeds 
-                }
-                ref_pred = ref_denoiser(**ref_inputs).sample
-                ref_diff, ref_loss = split_mse(ref_pred, target)
-        
-        # DPO Objective
-        surr_loss = -beta * (model_diff - ref_diff)
-        loss = -1 * F.logsigmoid(surr_loss.mean())
-
-        # Get approx accuracy as models probability of giving chosen over rejected
-        acc = (surr_loss > 0).sum().float() / len(surr_loss)
-        acc += 0.5 * (surr_loss == 0).sum().float() / len(surr_loss) # 50% for when both match
-
-        if accelerator is None:
-            loss.backward()
-        else:
-            accelerator.backward(loss)
-
-        return {
-            "loss" : loss.item(),
-            "diffusion_loss" : base_loss.item(),
-            "accuracy" : acc.item(),
-            "ref_deviation" : (ref_loss - base_loss) ** 2
-        }
